Refactored test classes
Now all classes test just one piece of the model and either merge or extraction instead of being inside a huge extrator/merger test class
